Move LoadTables to the DB
We are loading tables, we can do this directly on the DB.
